운영 및 통계 내역에서 전체 부서와 전체 사용자로 검색 가능하게 함.

+++ src/main/java/com/freeneo/survey/domain/BranchHistory.java
@@ -9,7 +9,7 @@ public class BranchHistory {
 	private String branchName;
 	private int sendCount;
 	private int respondentCount;
-	private float responseRatio;
+	private double responseRatio;
 	
 	public String getBranchName() {
 		return branchName;
@@ -30,10 +30,10 @@ public class BranchHistory {
 		this.respondentCount = respondentCount;
 	}
 
-	public float getResponseRatio() {
+	public double getResponseRatio() {
 		return responseRatio;
 	}
-	public void setResponseRatio(float responseRatio) {
+	public void setResponseRatio(double responseRatio) {
 		this.responseRatio = responseRatio;
 	}
 	@Override
